+++ drux/params.py
@@ -1,4 +1,4 @@
 # -*- coding: utf-8 -*-
 """Drux parameters and constants."""
 
-DRUX_VERSION = "0.1"
+DRUX_VERSION = "0.2"