+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Changed
 - `README.md` modified
+- Release range bug in `time_for_release` fixed
 ## [0.1] - 2025-07-27
 ### Added
 - Base model