+++ CHANGELOG.md
@@ -9,7 +9,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Zero-order model
 ### Changed
 - `README.md` modified
+- Test system modified
 - Release range bug in `time_for_release` fixed
+- `matplotlib` import bug fixed
 ## [0.1] - 2025-07-27
 ### Added
 - Base model