+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [0.2] - 2025-09-27
 ### Added
 - Zero-order model
 - First-order model
@@ -20,5 +21,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Higuchi model
 
 
-[Unreleased]: https://github.com/openscilab/drux/compare/v0.1...dev
+[Unreleased]: https://github.com/openscilab/drux/compare/v0.2...dev
+[0.2]: https://github.com/openscilab/drux/compare/v0.1...v0.2
 [0.1]: https://github.com/openscilab/drux/compare/48548f0...v0.1